Added loading and completed UI
