+++ src/icons/Checkmark.tsx
@@ -1,0 +1,16 @@
+export default function () {
+  return (
+    <svg
+      width="16"
+      height="11"
+      viewBox="0 0 16 11"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M1.75 7L3.83579 9.08579C4.61684 9.86684 5.88317 9.86683 6.66421 9.08579L14.25 1.5"
+        stroke="currentColor"
+        stroke-width="2"
+      />
+    </svg>
+  )
+}
